restruct tests
