+++ task2/solution.py
@@ -18,12 +18,13 @@ async def get_count_of_animals(soup: BeautifulSoup) -> dict[str, int]:
     all_groups = mw_pages.find_all('div', class_='mw-category-group')
     for group in all_groups:
         main_char = group.find('h3').text
-        animals_links = group.find_all('a', href=re.compile(f'{main_char}*'))
+        print(main_char)
+        animals_links = group.find_all('a', string=re.compile(f'{main_char}.*'))
         animals_counter[main_char] = len(animals_links)
     return animals_counter
 
 
-async def get_next_page(soup: BeautifulSoup) -> str:
+async def get_next_page(soup: BeautifulSoup) -> str | None:
     try:
         mw_pages = soup.find('div', id='mw-pages')
         next_page = mw_pages.find('a', string=re.compile('Следующая страница')).get('href')
@@ -41,7 +42,7 @@ def write_data(data: list):
 
 async def main() -> None:
     async with ClientSession(connector=TCPConnector(ssl=False)) as session:
-        animals_counter = Counter()
+        animals_counter: Counter = Counter()
         bas_url = 'https://ru.wikipedia.org/'
         url = 'https://ru.wikipedia.org/wiki/Категория:Животные_по_алфавиту'
         while True:
@@ -53,32 +54,10 @@ async def main() -> None:
             if not next_page:
                 break
             url = bas_url + next_page
+        print(animals_counter.total())
         char_animal_list = sorted(animals_counter.items())
         write_data(char_animal_list)
 
 
-async def test_count_animals_on_page():
-    async with ClientSession(connector=TCPConnector(ssl=False)) as session:
-        animals_counter = Counter()
-        url = 'https://ru.wikipedia.org/wiki/Категория:Животные_по_алфавиту'
-        page = await get_page(session, url)
-        soup = BeautifulSoup(page, 'html.parser')
-        count_of_animals = await get_count_of_animals(soup)
-        animals_counter.update(count_of_animals)
-
-        assert animals_counter.total() == 200
-
-        bas_url = 'https://ru.wikipedia.org/'
-        next_page = await get_next_page(soup)
-        url = bas_url + next_page
-        page = await get_page(session, url)
-        soup = BeautifulSoup(page, 'html.parser')
-        count_of_animals = await get_count_of_animals(soup)
-        animals_counter.update(count_of_animals)
-
-        assert animals_counter.total() == 400
-
-
 if __name__ == '__main__':
-    asyncio.run(test_count_animals_on_page())
     asyncio.run(main())
